+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Removed uuid array
 - Modified MLA citation rules
 - Deduplicated MLA citation generator
+- Replaced polyglot with fasttext for language detection
 
 ## 2022-01-25 -- v0.9.6
 ### Fixed